requirements.txt
protobuf==4.24.0
python-memcached==1.59
six==1.16.0
